2973045: Export to excel results in strange rounding; Thanks to Bogdan Cristian Paulon for the patch

+++ sql12/fw/src/net/sourceforge/squirrel_sql/fw/gui/action/TableExportCsvCommand.java
@@ -468,7 +468,7 @@ public class TableExportCsvCommand
             ret = new jxl.write.Number(colIdx, curRow, ((Number) cellObj).floatValue());
             break;
          case Types.NUMERIC:
-            ret = new jxl.write.Number(colIdx, curRow, ((Number) cellObj).floatValue());
+            ret = new jxl.write.Number(colIdx, curRow, ((Number) cellObj).doubleValue());
             break;
          case Types.FLOAT:
             ret = new jxl.write.Number(colIdx, curRow, ((Number) cellObj).floatValue());
